Consolidates config template card component
Move FileSourceTemplateCard to the generic SourceTemplateCard component under ConfigTemplates
Unifies template selection handling for file sources and object stores
